+++ example/lib/swagger_generated_code/pet_service_yaml.swagger.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: type=lint
 
 import 'package:json_annotation/json_annotation.dart';
+import 'package:json_annotation/json_annotation.dart' as json;
 import 'package:collection/collection.dart';
 import 'dart:convert';
 
@@ -30,7 +31,7 @@ abstract class PetServiceYaml extends ChopperService {
     ErrorConverter? errorConverter,
     Converter? converter,
     Uri? baseUrl,
-    Iterable<dynamic>? interceptors,
+    List<Interceptor>? interceptors,
   }) {
     if (client != null) {
       return _$PetServiceYaml(client);
@@ -108,6 +109,7 @@ abstract class PetServiceYaml extends ChopperService {
 
   ///Finds Pets by tags
   ///@param tags Tags to filter by
+  @deprecated
   Future<chopper.Response<List<Pet>>> petFindByTagsGet({
     required List<String>? tags,
     dynamic cacheControl,
@@ -120,6 +122,7 @@ abstract class PetServiceYaml extends ChopperService {
 
   ///Finds Pets by tags
   ///@param tags Tags to filter by
+  @deprecated
   @Get(path: '/pet/findByTags')
   Future<chopper.Response<List<Pet>>> _petFindByTagsGet({
     @Query('tags') required List<String>? tags,
@@ -502,7 +505,7 @@ class Order {
   static const fromJsonFactory = _$OrderFromJson;
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other is Order &&
             (identical(other.id, id) ||
@@ -590,7 +593,7 @@ class Category {
   static const fromJsonFactory = _$CategoryFromJson;
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other is Category &&
             (identical(other.id, id) ||
@@ -658,7 +661,7 @@ class User {
   static const fromJsonFactory = _$UserFromJson;
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other is User &&
             (identical(other.id, id) ||
@@ -761,7 +764,7 @@ class Tag {
   static const fromJsonFactory = _$TagFromJson;
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other is Tag &&
             (identical(other.id, id) ||
@@ -828,7 +831,7 @@ class Pet {
   static const fromJsonFactory = _$PetFromJson;
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other is Pet &&
             (identical(other.id, id) ||
@@ -918,7 +921,7 @@ class ApiResponse {
   static const fromJsonFactory = _$ApiResponseFromJson;
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other is ApiResponse &&
             (identical(other.code, code) ||
@@ -985,8 +988,9 @@ enums.OrderStatus? orderStatusNullableFromJson(
   if (orderStatus == null) {
     return null;
   }
-  return enums.OrderStatus.values
-          .firstWhereOrNull((e) => e.value == orderStatus) ??
+  return enums.OrderStatus.values.firstWhereOrNull((e) =>
+          e.value.toString().toLowerCase() ==
+          orderStatus.toString().toLowerCase()) ??
       defaultValue;
 }
 
@@ -1050,7 +1054,9 @@ enums.PetStatus? petStatusNullableFromJson(
   if (petStatus == null) {
     return null;
   }
-  return enums.PetStatus.values.firstWhereOrNull((e) => e.value == petStatus) ??
+  return enums.PetStatus.values.firstWhereOrNull((e) =>
+          e.value.toString().toLowerCase() ==
+          petStatus.toString().toLowerCase()) ??
       defaultValue;
 }
 
@@ -1116,8 +1122,9 @@ enums.PetFindByStatusGetStatus? petFindByStatusGetStatusNullableFromJson(
   if (petFindByStatusGetStatus == null) {
     return null;
   }
-  return enums.PetFindByStatusGetStatus.values
-          .firstWhereOrNull((e) => e.value == petFindByStatusGetStatus) ??
+  return enums.PetFindByStatusGetStatus.values.firstWhereOrNull((e) =>
+          e.value.toString().toLowerCase() ==
+          petFindByStatusGetStatus.toString().toLowerCase()) ??
       defaultValue;
 }