+++ lib/src/code_generators/swagger_enums_generator.dart
@@ -156,17 +156,20 @@ ${allEnums.map((e) => e.toString()).join('\n')}
     swaggerRoot.paths.forEach((String path, SwaggerPath swaggerPath) {
       swaggerPath.requests
           .forEach((String requestType, SwaggerRequest swaggerRequest) {
-        final successResponse = SwaggerRequestsGenerator.getSuccessedResponse(
+        final successResponses = SwaggerRequestsGenerator.getSuccessedResponses(
             responses: swaggerRequest.responses);
-        final successResponseSchema =
-            successResponse?.schema ?? successResponse?.content?.schema;
 
-        if (successResponseSchema != null) {
-          final responseEnums = generateEnumsFromSchemaMap({
-            '${path.pascalCase}${requestType.pascalCase}\$$kResponse':
-                successResponseSchema
-          });
-          result.addAll(responseEnums);
+        for (final successResponse in successResponses) {
+          final successResponseSchema =
+              successResponse.schema ?? successResponse.content?.schema;
+
+          if (successResponseSchema != null) {
+            final responseEnums = generateEnumsFromSchemaMap({
+              '${path.pascalCase}${requestType.pascalCase}\$$kResponse':
+                  successResponseSchema
+            });
+            result.addAll(responseEnums);
+          }
         }
 
         final parameters = [
@@ -350,9 +353,9 @@ ${allEnums.map((e) => e.toString()).join('\n')}
       );
 
       if (propertiesContainer != null) {
-        properties = propertiesContainer.properties;
+        properties = Map.from(propertiesContainer.properties);
       } else {
-        properties = schema.properties;
+        properties = Map.from(schema.properties);
       }
 
       var allOfRef = allOf.firstWhereOrNull((e) => e.hasRef);